django/auth/requirements.txt
aniso8601==9.0.1
apturl==0.5.2
asn1crypto==0.24.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
httplib2==0.14.0
idna==2.6
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
minepy==1.2.5
monotonic==1.5
mysql-connector-python-rf==2.2.2
netifaces==0.10.4
numpy==1.20.3
oauthlib==3.1.0
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.1.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.4.15
systemd-python==234
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==2.0.1
xkit==0.0.0
